+++ bokehjs/test/unit/models/glyphs/rect.ts
@@ -112,8 +112,8 @@ describe("Rect", () => {
       const glyph_view = await create_glyph_view(glyph, data, {axis_type: "linear"})
       glyph_view.map_data()
 
-      expect(glyph_view.sw).to.be.equal(Float64Array.of(20))
-      expect(glyph_view.sh).to.be.equal(Float64Array.of(40))
+      expect(glyph_view.sw).to.be.equal(Float32Array.of(20))
+      expect(glyph_view.sh).to.be.equal(Float32Array.of(40))
     })
 
     it("`_map_data` should correctly map data if width and height units are 'screen'", async () => {
@@ -124,8 +124,8 @@ describe("Rect", () => {
       const glyph_view = await create_glyph_view(glyph, data, {axis_type: "linear"})
       glyph_view.map_data()
 
-      expect(glyph_view.sw).to.be.equal(new Float64Array([10]))
-      expect(glyph_view.sh).to.be.equal(new Float64Array([20]))
+      expect(glyph_view.sw).to.be.equal(new Float32Array([10]))
+      expect(glyph_view.sh).to.be.equal(new Float32Array([20]))
     })
 
     // XXX: needs update
@@ -134,8 +134,8 @@ describe("Rect", () => {
       const glyph_view = await create_glyph_view(glyph, data)
       glyph_view.map_data()
 
-      expect(glyph_view.sx0).to.be.equal(Float64Array.of(0))
-      expect(glyph_view.sy1).to.be.equal(Float64Array.of(0))
+      expect(glyph_view.sx0).to.be.equal(Float32Array.of(0))
+      expect(glyph_view.sy1).to.be.equal(Float32Array.of(0))
     })
 
     // XXX: needs update
@@ -147,8 +147,8 @@ describe("Rect", () => {
       const glyph_view = await create_glyph_view(glyph, data)
       glyph_view.map_data()
 
-      expect(glyph_view.sx0).to.be.equal(Float64Array.of(-5))
-      expect(glyph_view.sy1).to.be.equal(Float64Array.of(-10))
+      expect(glyph_view.sx0).to.be.equal(Float32Array.of(-5))
+      expect(glyph_view.sy1).to.be.equal(Float32Array.of(-10))
     })
 
     /* XXX
@@ -179,8 +179,8 @@ describe("Rect", () => {
       const glyph_view = await create_glyph_view(glyph, data, {axis_type: "linear"})
       glyph_view.map_data()
 
-      expect(glyph_view.sw).to.be.equal(Float64Array.of(20))
-      expect(glyph_view.sh).to.be.equal(Float64Array.of(0))
+      expect(glyph_view.sw).to.be.equal(Float32Array.of(20))
+      expect(glyph_view.sh).to.be.equal(Float32Array.of(0))
     })
 
     describe("hit-testing", () => {
